+++ smelter-api/src/output/rtmp.rs
@@ -41,7 +41,7 @@ pub enum RtmpClientVideoEncoderOptions {
         bitrate: Option<VideoEncoderBitrate>,
 
         /// (**default=`5000`**) Interval between keyframes, in milliseconds.
-        keyframe_interval: Option<u64>,
+        keyframe_interval: Option<f64>,
 
         /// (**default=`"yuv420p"`**) Encoder pixel format
         pixel_format: Option<PixelFormat>,
@@ -56,7 +56,7 @@ pub enum RtmpClientVideoEncoderOptions {
         bitrate: Option<VideoEncoderBitrate>,
 
         /// (**default=`5000`**) Interval between keyframes, in milliseconds.
-        keyframe_interval: Option<u64>,
+        keyframe_interval: Option<f64>,
     },
 }
